requirements.txt
streamlit
pdfplumber
pytesseract
pillow
pandas
networkx
pyvis
spacy
